+++ lib/contextual_menu.dart
@@ -3,24 +3,52 @@ import 'package:popup_menu_2/popup_menu.dart';
 import 'package:popup_menu_2/popup_menu_item.dart';
 
 class ContextualMenu extends StatefulWidget {
+  /// the child is a [Widget]
   final Widget child;
+
+  /// list of items
   final List<MenuItemProvider> items;
+
+  /// background of the pop up
   final Color? backgroundColor;
+
+  /// highlight  of the pop up selected item
   final Color? highlightColor;
+
+  /// color  of line separator
   final Color? lineColor;
-  final GlobalKey globalKey;
+
+  /// context of pop up
+  final BuildContext ctx;
+
+  /// the key of the targetting pop up
+  final GlobalKey targetWidgetKey;
+
+  /// called after dismissing the popup
   final Function()? onDismiss;
+
+  /// called after an item clicked
   final Function(bool)? stateChanged;
+
+  /// chose if dissmiss if user click away
+  final bool dismissOnClickAway;
+
+  /// max columns
+  final int maxColumns;
   const ContextualMenu({
-    required this.globalKey,
+    Key? key,
+    required this.targetWidgetKey,
     required this.child,
     required this.items,
+    required this.ctx,
     this.backgroundColor,
     this.highlightColor,
     this.lineColor,
     this.onDismiss,
+    this.dismissOnClickAway = true,
     this.stateChanged,
-  }) : super(key: globalKey);
+    this.maxColumns = 3,
+  }) : super(key: key);
 
   @override
   State<ContextualMenu> createState() => _ContextualMenuState();
@@ -33,16 +61,16 @@ class _ContextualMenuState extends State<ContextualMenu> {
       child: widget.child,
       onTap: () {
         var popupMenu = PopupMenu(
-          context: context,
+          context: widget.ctx,
           backgroundColor: widget.backgroundColor,
           lineColor: widget.lineColor,
-          maxColumn: 3,
+          maxColumns: 3,
           items: widget.items,
           highlightColor: widget.highlightColor,
           stateChanged: widget.stateChanged,
           onDismiss: widget.onDismiss,
         );
-        popupMenu.show(widgetKey: widget.globalKey);
+        popupMenu.show(widgetKey: widget.targetWidgetKey);
       },
     );
   }