+++ lib/popup_menu_2.dart
@@ -1,7 +1,5 @@
 library popup_menu_2;
 
 export 'contextual_menu.dart';
-export 'popup_menu_item_widget.dart';
 export 'popup_menu_item.dart';
 export 'popup_menu.dart';
-export 'triangle_painter.dart';