+++ CHANGELOG.md
@@ -1,3 +1,13 @@
 # a pupup menu, for porpose not break the user attention
 
 **v0.0.4**
+Initial release 
+
+**v0.0.5**
+fix error: two widgets using the same global key
+
+
+**v0.0.6**
+add dismiss on click away
+add waiting alternative when item clicked while waiting
+improving the code